src/requirements.txt
PyYAML==6.0.1
sanic==23.12.1
